Convert save data loading format to dynamic save data creation
It used to have save data set specifically for levels 1-6. In version
2.0, however, I needed to make it compatable with the format since the
save data would be the same for any custom level used. It now uses the
name of the level for the save file, while also categorizing the save
data in the following folder format "Creator/Difficulty/Mode/Title.txt".
I could probably restructure the code a little bit, but Id like to get
this in for version 2.0 as soon as i can.